Initial USB console support
This commit needs serious tidying up, as it's the end-result of a lot of experimentation. I want to get it into VCS first, though.

This adds a CDC-ACM callback for incoming data that feeds into the micropython input ringbuffer. It also patches out part of the ESP32 port's stdout implementation to only use USB rather than UART.

Known issues:
 - The ACM flow control signals are not triggering restart into bootloader mode, so you still need to use the boot button
 - The implementation is sloppy, I don't understand why I can't cause the port's code to write to USB - I suspect a tinyusb conflict
 - This requires re-running ./scripts/firstTime.sh to trigger re-patching of the underlying files