Add a simple boot menu on the LCD that's navigable using the joystick
Containing for the moment just the two options that were previously
hard-coded in boot.py - wrappers around joystick_active() and
web_repl.start()

Also moved button definitions into tidal.py and included the 8x8 bitmap
font from st7789 in the ROM.

I don't expect this UI to be particularly reusable, or necessarily ship
with the badge, but hopefully it will be useful during bring up. I'm
hoping it makes it easier for people to add mini-apps for testing
hardware, and similar.