Support screen and buttons rotation
Also added support for badge.team compatible buttons API. on_press and
on_up_down callbacks no longer take the pin as a first argument, because
it's difficult for that to make sense if the buttons are rotated and
nothing was using it anyway.

Pressing B on the app_launcher will now rotate the screen.

Fixed PagedApp so that there's a window set up by the time
PagedApp.on_activate() is called (which is what apps should be doing
and should be documented somewhere...)