Update anti-sleep method
This removes the simple check for usb connected, for reasons of performance and reliability. Instead, it offers a Python API for determining if it's safe to sleep by checking the USB state, WiFi state, and a timer that can be manually set. We should call the new inhibit_sleep() method when detecting a USB charge event to ensure that we wake from sleep and stay awake long enough for a host to send configuration if we're plugged into a computer. The actual interrupt isn't in place yet.

Also, this makes the devboard prototype never sleep, as that has no chargedet pin.